+++ packages/core/src/config.ts
@@ -1,11 +1,6 @@
 import fs from 'node:fs';
 import { isAbsolute, join } from 'node:path';
-import {
-  DEFAULT_ASSET_PREFIX,
-  RspackChain,
-  color,
-  isObject,
-} from '@rsbuild/shared';
+import { RspackChain, color } from '@rsbuild/shared';
 import type {
   InspectConfigOptions,
   NormalizedConfig,
@@ -24,6 +19,7 @@ import type {
 } from '@rsbuild/shared';
 import {
   CSS_DIST_DIR,
+  DEFAULT_ASSET_PREFIX,
   DEFAULT_DATA_URL_SIZE,
   DEFAULT_DEV_HOST,
   DEFAULT_MOUNT_ID,
@@ -45,6 +41,7 @@ import {
   findExists,
   getNodeEnv,
   isFileExists,
+  isObject,
   upperFirst,
 } from './helpers';
 import { logger } from './logger';
