+++ packages/core/src/pluginManager.ts
@@ -1,4 +1,4 @@
-import { color, isFunction } from '@rsbuild/shared';
+import { color } from '@rsbuild/shared';
 import type {
   BundlerPluginInstance,
   Falsy,
@@ -6,6 +6,7 @@ import type {
   RsbuildPlugin,
   RsbuildPluginAPI,
 } from '@rsbuild/shared';
+import { isFunction } from './helpers';
 import { logger } from './logger';
 
 function validatePlugin(plugin: unknown) {
