+++ packages/core/src/plugins/moduleFederation.ts
@@ -1,10 +1,7 @@
-import {
-  type CacheGroup,
-  DEFAULT_ASSET_PREFIX,
-  type Rspack,
-} from '@rsbuild/shared';
+import type { CacheGroup, Rspack } from '@rsbuild/shared';
 import { rspack } from '@rspack/core';
 import type { RspackPluginInstance } from '@rspack/core';
+import { DEFAULT_ASSET_PREFIX } from '../constants';
 import type { RsbuildPlugin } from '../types';
 
 /**
