+++ packages/core/src/plugins/output.ts
@@ -1,11 +1,14 @@
 import { posix } from 'node:path';
-import {
-  DEFAULT_ASSET_PREFIX,
-  type NormalizedEnvironmentConfig,
-  type RsbuildContext,
+import type {
+  NormalizedEnvironmentConfig,
+  RsbuildContext,
 } from '@rsbuild/shared';
 import { rspack } from '@rspack/core';
-import { DEFAULT_DEV_HOST, DEFAULT_PORT } from '../constants';
+import {
+  DEFAULT_ASSET_PREFIX,
+  DEFAULT_DEV_HOST,
+  DEFAULT_PORT,
+} from '../constants';
 import { formatPublicPath, getFilename } from '../helpers';
 import { getCssExtractPlugin } from '../pluginHelper';
 import type { RsbuildPlugin } from '../types';
